requirements.txt
python-dotenv==1.0.0
google-auth-oauthlib==1.2.0
google-api-python-client==2.108.0
azure-cosmos==4.5.1
openai==1.3.7
langgraph==0.0.15
pydantic==2.5.2
pytz==2023.3
numpy==1.26.2
